+++ SEN4992/Pages/MainView.swift
@@ -11,17 +11,17 @@ struct MainView: View {
     let userApi = UserApi()
     
     var body: some View {
-        VStack {
-            Text("Hello, World!\nuser id:  \(userApi.currentUserId)\nname: \(userApi.currentUserName)")
-            CustomButtonUI(function: {
-                userApi.logOut {
-                    userApi.userLoginPageCheck()
-                } onError: { err in
-                    print(err)
+                VStack {
+                    Text("Hello, World!\nuser id:  \(userApi.currentUserId)\nname: \(userApi.currentUserName)")
+                    CustomButtonUI(function: {
+                        userApi.logOut {
+                            userApi.userLoginPageCheck()
+                        } onError: { err in
+                            print(err)
+                        }
+        
+                    }, title: "LogOut")
                 }
-
-            }, title: "LogOut")
-        }
     }
 }
